Move compute_*.py into horovod.tensorflow.data, fix examples
