+++ examples/tensorflow2/tensorflow2_mnist_data_service_train_fn_training_side_dispatcher.py
@@ -15,25 +15,21 @@
 
 import os
 
-import horovod.tensorflow.keras as hvd
 import tensorflow as tf
 
-from nereus.tf_data_service import TfDataServiceConfig, tf_data_service
-from nereus.utils.logging_utils import initialise_default_logging
-
-initialise_default_logging(os.path.basename(__file__))
+import horovod.tensorflow.keras as hvd
+from horovod.tensorflow.data.compute_service import TfDataServiceConfig, tf_data_service
 
 
 # arguments reuse_dataset and round_robin only used when single dispatcher is present
-def train_fn(compute_config: TfDataServiceConfig, reuse_dataset: bool = False, round_robin: bool = False):
+def train_fn(dataset_path: str, compute_config: TfDataServiceConfig, reuse_dataset: bool = False, round_robin: bool = False):
     # Horovod: initialize Horovod.
     hvd.init()
     rank = hvd.rank()
     size = hvd.size()
 
     with tf_data_service(compute_config, rank) as dispatcher_address:
-        (mnist_images, mnist_labels), _ = \
-            tf.keras.datasets.mnist.load_data(path=f'{os.getcwd()}/mnist/mnist.npz')
+        (mnist_images, mnist_labels), _ = tf.keras.datasets.mnist.load_data(path=dataset_path)
 
         dataset = tf.data.Dataset.from_tensor_slices(
             (tf.cast(mnist_images[..., tf.newaxis] / 255.0, tf.float32),
